chore: Gnb에서 UserProfile 분리하고 에러 로직 선언적으로 개선 	modified:   components/layouts/header/gnb/Gnb.module.css 	modified:   components/layouts/header/gnb/Gnb.tsx 	new file:   components/layouts/header/gnb/UserProfile.module.css 	new file:   components/layouts/header/gnb/UserProfile.tsx
