+++ components/layouts/header/gnb/Gnb.module.css
@@ -10,21 +10,6 @@
   height: 1.6rem;
 }
 
-.gnb-profile {
-  gap: 0.6rem;
-}
-
-.profile-img {
-  width: 2.8rem;
-  height: 2.8rem;
-  border-radius: 9999rem;
-}
-
-.profile-email {
-  font-size: 1.4rem;
-  font-weight: 400;
-}
-
 .nav-dummy {
   height: 6.3rem;
 }
