+++ components/layouts/header/gnb/Gnb.tsx
@@ -3,36 +3,21 @@ import Image from 'next/image';
 import Link from 'next/link';
 import { useRouter } from 'next/router';
 
-import { LOADING_MESSAGE, ROUTE_PATHS } from 'constants/constants';
-import useFetch from 'hooks/useFetch';
+import { ROUTE_PATHS } from 'constants/constants';
 import scrollToTop from 'utils/scrollToTop';
 
-import { USER_API_URL, UserIdApiResponse } from '@/apis/api';
-import ErrorMessage from '@/components/common/ErrorMessage';
-import LoginButton from '@/components/common/buttons/LoginButton';
 import styles from '@/components/layouts/header/gnb/Gnb.module.css';
 import logoSvgUrl from '@/public/images/logo.svg';
-import defaultProfileImg from '@/public/images/profile-img.png';
+
+import UserProfile from './UserProfile';
 
 const containerClasses = classNames('flex-row', 'align-center', 'justify-space-between');
 const logoClasses = classNames(styles['gnb-logo'], 'position-relative');
-const profileClasses = classNames(styles['gnb-profile'], 'flex-row', 'align-center');
-const profileImgClasses = classNames(styles['profile-img'], 'position-relative');
-const profileEmailClasses = classNames(styles['profile-email'], 'hidden-block-mobile-only', 'text-color-gray100');
 
 // 글로벌 네비게이션 바
 function Gnb() {
-  const url = USER_API_URL;
-  const { data, loading, error } = useFetch<UserIdApiResponse>(url);
-
-  // data를 가공하여 userInfo에 저장
-  const userInfo = data?.data[0] ?? null;
-
   // 페이지 경로 저장
   const { pathname } = useRouter();
-
-  const userProfileImg = userInfo?.profileImageSource || defaultProfileImg;
-  const userEmail = userInfo?.email ?? '';
 
   // /folder에서는 position-fixed 제거
   const navClasses = classNames(
@@ -53,18 +38,7 @@ function Gnb() {
               <Image src={logoSvgUrl} alt="logo" fill />
             </div>
           </Link>
-          {loading && <ErrorMessage message={LOADING_MESSAGE} />}
-          {userInfo ? (
-            <div className={profileClasses}>
-              <div className={profileImgClasses}>
-                <Image src={userProfileImg} alt="profile-img" fill />
-              </div>
-              <p className={profileEmailClasses}>{userEmail}</p>
-            </div>
-          ) : (
-            <LoginButton />
-          )}
-          {error !== null && <ErrorMessage message={String(error)} />}
+          <UserProfile />
         </div>
       </nav>
       {/* 더미 요소로 공간 차지 */}
