+++ components/layouts/header/gnb/UserProfile.module.css
@@ -1,0 +1,14 @@
+.gnb-profile {
+  gap: 0.6rem;
+}
+
+.profile-img {
+  width: 2.8rem;
+  height: 2.8rem;
+  border-radius: 9999rem;
+}
+
+.profile-email {
+  font-size: 1.4rem;
+  font-weight: 400;
+}
